requirements.txt
asgiref==3.11.1
charset-normalizer==3.4.4
Django==6.0.2
django-rest-framework==0.1.0
djangorestframework==3.16.1
numpy==2.4.2
pandas==3.0.0
pillow==12.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
reportlab==4.4.9
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
